move dask code into xarray/backends/locks.py
